vfio: update for new iommu groups
This is a fairly minimal change.  Previously the iommu_group entry
in the device sysfs directory was a file that we read the iommu
group number from.  Now it's a link to the group, so we instead
need to do a readlink and strip just the group number from it.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>