vfio: Overhaul BAR mapping
First, these were called resources because PCI-sysfs had "resource"
files that KVM device assignment used.  Obviously we don't do that,
so why call them resources, they're BARs.

The main point of this patch though is that we can be a little more
creative with overlapping memory regions to build up a BAR.  All BARs
are now mapped slow.  If supported, we build on top of that subregions
that mmap the MMIO.  If the MSIX vector table occupies part of the BAR
mapping, we skip that section and later let msix_init add a subregion
over it.  VFIO doesn't care if we mmap the MSIX PBA, so we leave that
and let memory region code deal with exposing the emulated PBA on top
of it.

We also get rid of the silly MSIX_PAGE_SIZE here.  We have to align
the start and end of the vector table to page boundaries, but that's
really more a mapping artifact than anything specific to MSIX.  Since
we have a slow mapped memory region across the whole BAR, anything
in the gap between mmaped space and emulated vector table will still
be accessible using read/write.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>