vfio: Support MSI direct to KVM
KVM now supports a generic interface for routing an eventfd
directly into an irqfd so we can completely bypass qemu for
MSI & MSIX interrupts.  Qemu support is better for MSIX than
MSI, so we unfortunately have a hack to cobble together an
MSIMessage.  This also switches us to the MSIX vector_user
and vector_release callbacks, so MSIX is allocated as vectors
are added.  It's a bit ugly how we have to turn off MSIX and
re-enable it with more vectors when they're added, but there's
not much other choice.  MSI-X tested with 82576 VF, MSI with
Intel HDA audio.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>