vfio: Add support for non-upstream KVM_IRQFD_LEVEL_EOI
This ties the INTx eventfd directly into a KVM level irqfd.  Upon
receiving an EOI from the guest, KVM de-asserts the IOAPIC pin
and triggers an EOI eventfd.  This eventfd is tied directly into
an unmask irqfd in VFIO.  Thus we bypass userspace in both
directions of sending and acknowledging INTx.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>