vfio: Move iommu specific data to a union
This should help in adding support for other archs.  Create a
structure, add it to the union, populate it with map and unmap
function pointers, each of which take a pointer to your structure
as an argument, register that structure with something that will
call it when there's mappings to change.  If you need to free or
unregister, set the release function.  Map and unmap functions can
use container_of to get to the container and thefore the fd to use
to get to the iommu.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>